Make it more robust when very large memory blocks are used e.g. more than 2GB.  This happens when high samplerates and/or a high number of signals is present in  combination with long timescales.

+++ edfplusd_cnv.cpp
@@ -376,7 +376,7 @@ void UI_EDFDwindow::SelectFileButton()
       former_timestamp = next_timestamp - edfhdr->long_data_record_duration;
     }
 
-    if(next_timestamp!=(former_timestamp + edfhdr->long_data_record_duration))
+    if(next_timestamp != (long long)(former_timestamp + edfhdr->long_data_record_duration))
     {
       chunk_endtime = next_timestamp;
